infra/script/rag/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
azure-common==1.1.28
azure-core==1.36.0
azure-search-documents==11.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
jiter==0.12.0
openai==2.8.1
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pypdfium2==5.1.0
pytesseract==0.3.13
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
